% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_exercises.R
\docType{data}
\name{.exercise_db_base}
\alias{.exercise_db_base}
\title{Exercise information lookup}
\format{
An object of class \code{character} of length 1.
}
\usage{
.exercise_db_base
}
\description{
Data sourced from the free-exercise-db (github.com/yuhonas/free-exercise-db)
Images: https://raw.githubusercontent.com/yuhonas/free-exercise-db/main/exercises/{id}/{0|1}.jpg
}
\keyword{datasets}
